rename skia_c_bindings
